add contract tests

+++ packages/contracts/src/interfaces/IRsaVerifier.sol
@@ -3,5 +3,10 @@
 pragma solidity ^0.8.0;
 
 interface IRsaVerifier {
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[21] calldata _pubSignals) external view returns (bool);
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[21] calldata _pubSignals
+    ) external view returns (bool);
 }
